Revert "HB-44: MUI Login Page (#19)"
This reverts commit 1e03c22a7b78931ed884e834ba42a54320e3965e.
