+++ frontend/src/components/molecules/HamburgerMenu/HamburgerMenu.scss
@@ -1,11 +1,5 @@
 .HamburgerMenu {
   margin-top: 80px;
-  a:link,
-  a:visited,
-  a:hover,
-  a:active {
-    text-decoration: none;
-  }
   .MenuItem {
     display: flex;
     justify-content: center;
