--- frontend/src/components/molecules/UserSignInForm/UserSignInForm.scss
@@ -1,50 +1,0 @@
-.UserSignInForm {
-  padding: 0 10px;
-  display: flex;
-  flex-flow: column nowrap;
-  align-items: center;
-  justify-content: flex-start;
-  box-sizing: border-box;
-  height: 70%;
-  .usernameInput,
-  .passwordInput {
-    margin: 10px 0px;
-    width: 100%;
-  }
-
-  .credInput {
-    box-sizing: border-box;
-    width: 100%;
-  }
-
-  .formBttn {
-    height: 55px;
-    width: 100%;
-    color: #ffffff;
-    border-radius: 5px;
-    font-size: 6vw;
-    margin: 5px 0px;
-  }
-
-  .formBttns {
-    display: flex;
-    flex-flow: column nowrap;
-    margin-top: auto;
-    width: 100%;
-  }
-}
-
-@media screen and (min-width: 719px) {
-  .UserSignInForm {
-    width: 40%;
-    height: 350px;
-    .formBttn {
-      height: 55px;
-      width: 100%;
-      color: #ffffff;
-      border-radius: 5px;
-      font-size: 1.2rem;
-      margin: 5px 0px;
-    }
-  }
-}
