+++ frontend/src/components/molecules/UserSignUpForm/UserSignUpForm.css
@@ -1,0 +1,56 @@
+.username-input,
+.password-input,
+.email-input {
+  width: fit-content;
+  margin: 0 auto;
+  margin-bottom: 34px;
+}
+
+.username-input label,
+.password-input label,
+.email-input label {
+  width: fit-content;
+  display: block;
+}
+
+.username-input input,
+.password-input input,
+.email-input input {
+  height: 24px;
+  width: 200px;
+  border-top: none;
+  border-left: none;
+  border-right: none;
+  border-bottom: 1px solid#589bc1;
+}
+
+input[type="submit"] {
+  margin-bottom: 34px;
+}
+
+.username-input,
+.password-input,
+.email-input,
+.button {
+  color: antiquewhite;
+}
+
+.button2 {
+  font-size: 20px;
+  border-radius: 5px;
+  background-color: #313131;
+  border: 2px solid #9633c5;
+  color: antiquewhite;
+}
+
+.username-input input,
+.password-input input,
+.email-input input {
+  background-color: transparent;
+}
+
+.button {
+  font-size: 20px;
+  border-radius: 5px;
+  background-color: #589bc1;
+}
