+++ frontend/src/components/molecules/UserSignUpForm/UserSignUpForm.js
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import axios from "axios";
-import "./UserSignUpForm.scss";
+import "./UserSignUpForm.css";
 
 function UserSignupForm() {
   const API = process.env.REACT_APP_API_URL;
